improve alumni page
