+++ src/components/Alumni.astro
@@ -1,6 +1,5 @@
 ---
-import { currentMembers, alumniMembers } from "./data/alumni";
-import { Image } from "astro:assets";
+import { currentMembers, alumniMembers } from "./data/members";
 import { Headline } from "./Headline";
 import MemberCard from "./MemberCard.astro";
 ---
