+++ src/components/MemberCard.astro
@@ -6,16 +6,36 @@ const { member } = Astro.props as { member: Member };
 ---
 
 <div class="flex flex-col items-center justify-center w-24 m-4">
-    <Image
-        src={member.image}
-        alt={member.name}
-        quality="mid"
-        width={200}
-        height={200}
-        loading="lazy"
-        decoding="async"
-        class="rounded-full h-24 w-24 shadow-inner"
-    />
+    {member.image ? (
+        typeof member.image === 'string' ? (
+            <img
+                src={member.image}
+                alt={member.name}
+                width={200}
+                height={200}
+                loading="lazy"
+                decoding="async"
+                class="rounded-full h-24 w-24 shadow-inner"
+            />
+        ) : (
+            <Image
+                src={member.image}
+                alt={member.name}
+                quality="mid"
+                width={200}
+                height={200}
+                loading="lazy"
+                decoding="async"
+                class="rounded-full h-24 w-24 shadow-inner"
+            />
+        )
+    ) : (
+        <div class="rounded-full h-24 w-24 shadow-inner bg-neutral-900 flex items-center justify-center">
+            <span class="text-white text-2xl font-bold select-none">
+                {member.name.charAt(0)}{member.surname.charAt(0)}
+            </span>
+        </div>
+    )}
     <div class="text-center no-underline mt-2 text-secondary">
         {member.name}
         <br />
