+++ src/components/data/members.ts
@@ -21,13 +21,14 @@ import nina from "../../images/members/nina.png"
 import shairal from "../../images/members/shairal.png"
 import anna from "../../images/members/anna.png"
 import jan from "../../images/members/jan.png"
-
+import type { ImageMetadata } from "astro";
 
 export class Member {
-    constructor(public name: string, public surname: string, public image: string, public link: string | null) { }
+    constructor(public name: string, public surname: string, public image: ImageMetadata | string | null, public link: string | null) { }
 }
 
-export const members = [
+// Current members (active)
+export const currentMembers = [
     new Member("Justus", "Westerhoff", justus, "https://www.linkedin.com/in/justus-westerhoff/"),
     new Member("Cederic", "Aßmann", cederic, "https://www.linkedin.com/in/cederic-a%C3%9Fmann-41904322b"),
     new Member("Jonas", "Loos", jonas, "https://www.linkedin.com/in/jonas-loos/"),
@@ -47,12 +48,19 @@ export const members = [
     new Member("Nina", "Zukowska", nina, "https://www.linkedin.com/in/nina-zukowska/"),
     new Member("Shairal", "Sharma", shairal, "https://www.linkedin.com/in/shairal-om-sharma-409590188/"),
     new Member("Anna", "Tils", anna, "https://www.linkedin.com/in/anna-tils-0759b3312/"),
-    // new Member("Jan", "Tiegges", jan, "https://www.linkedin.com/in/jan-tiegges/"),
-    // new Member("Maja", "Kosiarski", null, "https://www.linkedin.com/in/maja-kosiarski-227194297/"),
-    // new Member("Mia", "Prokopovych ", mia, "https://www.linkedin.com/in/mia-prokopovych-9aaa26302"),
-    // new Member("Jarek", "Liesen", null, "https://www.linkedin.com/in/jarek-liesen/"),
-    // new Member("Josephine", "Egerer", josi, "https://www.linkedin.com/in/josephine-egerer/"),
-    // new Member("Jakob", "Hackstein", null, "https://www.linkedin.com/in/jakob-hackstein-5a1167281/"),
-    // new Member("Felix", "Ringe", null, "https://www.linkedin.com/in/felix-ringe/"),
-    // new Member("Philippa", "Ringe", philipa, null),
+    new Member("Maja", "Kosiarski", null, "https://www.linkedin.com/in/maja-kosiarski-227194297/"),
 ]
+
+// Alumni members (past members)
+export const alumniMembers = [
+    new Member("Jan", "Tiegges", jan, "https://www.linkedin.com/in/jan-tiegges/"),
+    new Member("Mia", "Prokopovych", mia, "https://www.linkedin.com/in/mia-prokopovych-9aaa26302"),
+    new Member("Jarek", "Liesen", null, "https://www.linkedin.com/in/jarek-liesen/"),
+    new Member("Josephine", "Egerer", josi, "https://www.linkedin.com/in/josephine-egerer/"),
+    new Member("Jakob", "Hackstein", null, "https://www.linkedin.com/in/jakob-hackstein-5a1167281/"),
+    new Member("Felix", "Ringe", null, "https://www.linkedin.com/in/felix-ringe/"),
+    new Member("Philippa", "Ringe", philipa, null),
+]
+
+// Legacy export for backward compatibility (current members only)
+export const members = currentMembers
